+++ CHANGELOG.md
@@ -56,6 +56,8 @@
   [#4099](https://github.com/Kong/kong-operator/pull/4099)
 - Hybridgateway: add support for `konghq.com/connect-timeout` service annotation
   [#4101](https://github.com/Kong/kong-operator/pull/4101)
+- Hybridgateway: add support for `konghq.com/read-timeout` service annotation
+  [#4102](https://github.com/Kong/kong-operator/pull/4102)
 - Add the following headers in requests of `ingress-controller` sent to Konnect
   for uploading configuration for tracing:
   - `X-Kic-Konnect-Sync-Instance-Id` for instance ID of Konnect config synchronizer.